refactor: extract useDashletPgrest and useHybridPgrestContext hooks
Reduce dashlet.tsx duplication by consolidating pgrest wiring:
- useDashletPgrest: wraps usePgrestResolvedFields with config defaults (8 files)
- useHybridPgrestContext: wraps usePgrestRows + data provider context merge (3 files)
- Export PgrestResolvedFieldsResult to fix TS4058
