+++ apps/app/src/features/dashboard/dashlets/stat_gradient/dashlet.tsx
@@ -3,7 +3,7 @@
 import { useMemo } from "react";
 import type { DashletComponentProps, DashletLayoutDefaults } from "../types";
 import type { PgrestParam, PgrestHttpMethod } from "../common";
-import { usePgrestResolvedFields, EMPTY_PGREST_PARAMS, DashletLoading, DashletError, parseResolvedNumber } from "../common";
+import { useDashletPgrest, DashletLoading, DashletError, parseResolvedNumber } from "../common";
 
 // ============================================================================
 // Configuration Types
@@ -66,14 +66,7 @@ export function Dashlet({ widget }: Readonly<DashletComponentProps>) {
     [config.title, config.value, config.unit],
   );
 
-  const { resolved, loading, fetchError } = usePgrestResolvedFields({
-    dataMode: (config.dataMode as "static" | "pgrest") || "static",
-    pgrestFunctionName: config.pgrestFunctionName || "",
-    pgrestHttpMethod: config.pgrestHttpMethod || "POST",
-    pgrestParams: config.pgrestParams || EMPTY_PGREST_PARAMS,
-    fields,
-    dataSourceId: config.dataSourceId,
-  });
+  const { resolved, loading, fetchError } = useDashletPgrest(config, fields);
 
   if (loading) return <DashletLoading />;
   if (fetchError) return <DashletError message={fetchError} />;
